Former-commit-id: 9c7d9ff3ad287f82a5772b00e2337e7edd24121d
Former-commit-id: b1ab56b05621a52d60c5799ea8299e20abaab9d8

+++ src/js/projects.js
@@ -3,7 +3,8 @@
 var _ = require('lodash'),
     path = require('path'),
     utils = require('../js/lib/utils'),
-    fs = require('fs');
+    fs = require('fs'),
+    trash = require('trash');
 
 function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
 
@@ -11,7 +12,6 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
         write = utils.fs.outputFile,
         read = utils.fs.readFile,
         mkdirp = utils.fs.mkdirs,
-        rm = utils.fs.remove,
         readdir = utils.fs.readdir,
         map = utils.lodash.map,
         flatten = utils.lodash.flatten,
@@ -21,12 +21,23 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
     return {
 
         moveBackups: function(oldPath, newPath) {
-            utils.fs.move(path.join(oldPath, 'automatic_backups'), path.join(newPath, 'automatic_backups'), {clobber: true});
-            utils.fs.move(path.join(oldPath, 'backups'), path.join(newPath, 'backups'), {clobber: true});
+            return utils.fs.mkdirs(configurator.getUserPath('datalocation', 'automatic_backups'))
+                .then(function () {
+                    return utils.fs.mkdirs(configurator.getUserPath('datalocation', 'backups'));
+                })
+                .then(function () {
+                    return utils.fs.stat(oldPath).then(App.utils.ret(true)).catch(App.utils.ret(false));
+                })
+                .then(function (exists) {
+                    if (exists) {
+                        utils.fs.move(path.join(oldPath, 'automatic_backups'), path.join(newPath, 'automatic_backups'), {clobber: true});
+                        utils.fs.move(path.join(oldPath, 'backups'), path.join(newPath, 'backups'), {clobber: true});
+                    }
+                });
         },
 
         updateManifestToMeta: function (manifest) {
-            var meta = manifest;
+            var meta = _.cloneDeep(manifest);
             try {
                 if (manifest.project.name === "") {
                     meta.project.name = dataManager.getProjectName(manifest.project.id)[0].name;
@@ -42,6 +53,7 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
                     meta.source_translations[j].id = details.id;
                     meta.source_translations[j].language_name = details.language_name;
                     meta.source_translations[j].resource_name = details.resource_name;
+                    meta.source_translations[j].direction = details.direction;
                 }
 
                 if (manifest.source_translations.length) {
@@ -58,18 +70,23 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
                     meta.project_type_class = "standard";
                 }
 
-                meta.unique_id = manifest.target_language.id + "_" + manifest.project.id + "_" + manifest.type.id;
-                if (manifest.resource.id !== "") {
-                    meta.unique_id += "_" + manifest.resource.id;
+                meta.unique_id = this.makeUniqueId(manifest);
+
+                if (!manifest.finished_chunks) {
+                    meta.finished_chunks = [];
                 }
 
                 var completion = configurator.getValue(meta.unique_id + "-completion");
                 if (completion !== undefined && completion !== "") {
                     meta.completion = completion;
                 } else {
-                    if (manifest.source_translations.length) {
+                    if (manifest.source_translations.length && manifest.finished_chunks) {
                         var frames = dataManager.getSourceFrames(manifest.source_translations[0]);
-                        meta.completion = Math.round((meta.finished_chunks.length / frames.length) * 100);
+                        if (frames.length) {
+                            meta.completion = Math.round((meta.finished_chunks.length / frames.length) * 100);
+                        } else {
+                            meta.completion = 0;
+                        }
                     } else {
                         meta.completion = 0;
                     }
@@ -81,35 +98,57 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
             return meta;
         },
 
-        saveTargetChunk: function (chunk, meta) {
+        makeUniqueId: function (manifest) {
+            var id = manifest.target_language.id + "_" + manifest.project.id + "_" + manifest.type.id;
+            if (manifest.resource.id !== "") {
+                id += "_" + manifest.resource.id;
+            }
+            return id;
+        },
+
+        updateChunk: function (meta, chunk) {
             var paths = utils.makeProjectPaths(targetDir, meta);
             var projectClass = meta.project_type_class;
+            var file = path.join(paths.projectDir, chunk.chunkmeta.chapterid, chunk.chunkmeta.frameid + '.txt');
+            var standardcontent = chunk.transcontent;
+            var hasContent = false;
 
-            var makeChapterDir = function (chunk) {
-                return mkdirp(path.join(paths.projectDir, chunk.chunkmeta.chapterid));
-            };
+            if (projectClass === "standard") {
+                hasContent = !!chunk.transcontent;
+            }
+            if (projectClass === "helps") {
+                hasContent = !!chunk.helpscontent.length;
+            }
+            if (projectClass === "extant" && chunk.helpscontent[0] && (!!chunk.helpscontent[0].title || !!chunk.helpscontent[0].body)) {
+                hasContent = true;
+            }
+            if (projectClass === "standard" && hasContent && chunk.chunkmeta.frame === 1 && chunk.projectmeta.project.id !== "obs") {
+                standardcontent = "\\c " + chunk.chunkmeta.chapter + " " + standardcontent;
+            }
+            return hasContent ? write(file, projectClass === "standard" ? standardcontent : toJSON(chunk.helpscontent)) : trash([file]);
+        },
 
-            var updateChunk = function (chunk) {
-                var file = path.join(paths.projectDir, chunk.chunkmeta.chapterid, chunk.chunkmeta.frameid + '.txt');
-                var hasContent = false;
-                if (projectClass === "standard") {
-                    hasContent = !!chunk.transcontent;
-                }
-                if (projectClass === "helps") {
-                    hasContent = !!chunk.helpscontent.length;
-                }
-                if (projectClass === "extant" && (!!chunk.helpscontent[0].title || !!chunk.helpscontent[0].body)) {
-                    hasContent = true;
-                }
-                return hasContent ? write(file, projectClass === "standard" ? chunk.transcontent : toJSON(chunk.helpscontent)) : rm(file);
-            };
+        makeChapterDir: function (meta, chunk) {
+            var paths = utils.makeProjectPaths(targetDir, meta);
 
-            return makeChapterDir(chunk)
-                .then(updateChunk(chunk));
+            return mkdirp(path.join(paths.projectDir, chunk.chunkmeta.chapterid));
+        },
+
+        saveTargetChunk: function (chunk, meta) {
+            var mythis = this;
+            return mythis.makeChapterDir(meta, chunk)
+                .then(function () {
+                    return mythis.updateChunk(meta, chunk);
+                })
+                .catch(function (err) {
+                    reporter.logError(err);
+                    throw "Unable to write to chunk file.";
+                });
         },
 
         saveTargetManifest: function (meta) {
             var paths = utils.makeProjectPaths(targetDir, meta);
+            var build = configurator.getAppData().build;
 
             var sources = meta.source_translations.map(function (source) {
                 return {
@@ -126,7 +165,7 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
                 format: meta.format,
                 generator: {
                     name: 'ts-desktop',
-                    build: ''
+                    build: build
                 },
                 target_language: meta.target_language,
                 project: meta.project,
@@ -138,84 +177,23 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
                 finished_chunks: meta.finished_chunks
             };
 
-            return write(paths.manifest, toJSON(manifest));
+            return write(paths.manifest, toJSON(manifest))
+                .catch(function (err) {
+                    reporter.logError(err);
+                    throw "Unable to write to manifest file.";
+                });
         },
 
-        saveTargetTranslation: function (translation, meta, user) {
+        createTargetTranslation: function (translation, meta, user) {
+            var mythis = this;
             var paths = utils.makeProjectPaths(targetDir, meta);
-            var projectClass = meta.project_type_class;
-            var sources = meta.source_translations.map(function (source) {
-                    return {
-                        language_id: source.language_id,
-                        resource_id: source.resource_id,
-                        checking_level: source.checking_level,
-                        date_modified: source.date_modified,
-                        version: source.version
-                    };
-                });
-
-            var manifest = {
-                package_version: meta.package_version,
-                format: meta.format,
-                generator: {
-                    name: 'ts-desktop',
-                    build: ''
-                },
-                target_language: meta.target_language,
-                project: meta.project,
-                type: meta.type,
-                resource: meta.resource,
-                source_translations: sources,
-                parent_draft: meta.parent_draft,
-                translators: meta.translators,
-                finished_chunks: meta.finished_chunks
-            };
-
-            var writeFile = function (name, data) {
-                return function () {
-                    return write(name, toJSON(data));
-                };
-            };
-
-            var makeChapterDir = function (chunk) {
-                return mkdirp(path.join(paths.projectDir, chunk.chunkmeta.chapterid));
-            };
+            var makeChapterDir = mythis.makeChapterDir.bind(this, meta);
+            var updateChunk = mythis.updateChunk.bind(this, meta);
 
             var makeChapterDirs = function (data) {
                 return function () {
                     return Promise.all(_.map(data, makeChapterDir));
                 };
-            };
-
-            var cleanChapterDir = function (data, chapter) {
-                var chapterpath = path.join(paths.projectDir, chapter);
-                return readdir(chapterpath).then(function (dir) {
-                    return !dir.length ? rm(chapterpath): true;
-                });
-            };
-
-            var cleanChapterDirs = function () {
-                return function () {
-                    var data = _.groupBy(translation, function (chunk) {
-                        return chunk.chunkmeta.chapterid;
-                    });
-                    return Promise.all(_.map(data, cleanChapterDir));
-                };
-            };
-
-            var updateChunk = function (chunk) {
-                var file = path.join(paths.projectDir, chunk.chunkmeta.chapterid, chunk.chunkmeta.frameid + '.txt');
-                var hasContent = false;
-                if (projectClass === "standard") {
-                    hasContent = !!chunk.transcontent;
-                }
-                if (projectClass === "helps") {
-                    hasContent = !!chunk.helpscontent.length;
-                }
-                if (projectClass === "extant" && (!!chunk.helpscontent[0].title || !!chunk.helpscontent[0].body)) {
-                    hasContent = true;
-                }
-                return hasContent ? write(file, projectClass === "standard" ? chunk.transcontent : toJSON(chunk.helpscontent)) : rm(file);
             };
 
             var updateChunks = function (data) {
@@ -226,7 +204,8 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
 
             var setLicense = function () {
                 var srcDir = path.resolve(path.join(__dirname, '..'));
-                return read(path.join(srcDir, 'assets', 'LICENSE.md'))
+                var file = meta.project.id === 'obs' ? 'OBS_LICENSE.md' : 'LICENSE.md';
+                return read(path.join(srcDir, 'assets', file))
                     .then(function(data) {
                         return write(paths.license, data);
                     });
@@ -234,15 +213,48 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
 
             return mkdirp(paths.projectDir)
                 .then(setLicense())
-                .then(writeFile(paths.manifest, manifest))
+                .then(function () {
+                    return mythis.saveTargetManifest(meta);
+                })
                 .then(makeChapterDirs(translation))
                 .then(updateChunks(translation))
-                .then(cleanChapterDirs())
                 .then(function () {
-                    return git.init(paths.projectDir);
+                    return mythis.cleanProject(translation, meta);
                 })
                 .then(function () {
-                    return git.stage(user, paths.projectDir);
+                    return mythis.commitProject(meta, user);
+                })
+                .catch(function (err) {
+                    throw "Error creating new project: " + err;
+                });
+        },
+
+        cleanProject: function (translation, meta) {
+            var paths = utils.makeProjectPaths(targetDir, meta);
+
+            var cleanChapterDir = function (data, chapter) {
+                var chapterpath = path.join(paths.projectDir, chapter);
+                return readdir(chapterpath).then(function (dir) {
+                    return !dir.length ? trash([chapterpath]): true;
+                }).catch(utils.ret(true));
+            };
+
+            var cleanChapterDirs = function () {
+                var data = _.groupBy(translation, function (chunk) {
+                    return chunk.chunkmeta.chapterid;
+                });
+                return Promise.all(_.map(data, cleanChapterDir));
+            };
+
+            return cleanChapterDirs();
+        },
+
+        commitProject: function (meta, user) {
+            var paths = utils.makeProjectPaths(targetDir, meta);
+
+            return git.init(paths.projectDir)
+                .then(function () {
+                    return git.commitAll(user, paths.projectDir);
                 });
         },
 
@@ -344,9 +356,31 @@ function ProjectsManager(dataManager, configurator, reporter, git, migrator) {
                 .then(utils.lodash.indexBy('name'));
         },
 
+        unsetValues: function (meta) {
+            var key = meta.unique_id;
+
+            configurator.unsetValue(key + "-chapter");
+            configurator.unsetValue(key + "-index");
+            configurator.unsetValue(key + "-selected");
+            configurator.unsetValue(key + "-completion");
+            configurator.unsetValue(key + "-source");
+        },
+
         deleteTargetTranslation: function (meta) {
             var paths = utils.makeProjectPaths(targetDir, meta);
-            return rm(paths.projectDir);
+
+            return utils.fs.stat(paths.projectDir).then(App.utils.ret(true)).catch(App.utils.ret(false))
+                .then(function (exists) {
+                    if (exists) {
+                        return trash([paths.projectDir]);
+                    } else {
+                        throw "Project file does not exist";
+                    }
+                })
+                .catch(function (err) {
+                    reporter.logError(err);
+                    throw "Unable to delete file at this time. You may need to restart the app first.";
+                });
         }
     };
 }
